Refactor Main App
