+++ lib/app.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:personal_finance_management_app/app/app.router.dart';
+import 'package:stacked_services/stacked_services.dart';
+
+class PersonalFinanceManagementApp extends StatelessWidget {
+  const PersonalFinanceManagementApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      navigatorKey: StackedService.navigatorKey,
+      onGenerateRoute: StackedRouter().onGenerateRoute,
+    );
+  }
+}
