+++ lib/main.dart
@@ -1,28 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:logger/logger.dart';
 import 'package:personal_finance_management_app/app/app.locator.dart';
-import 'package:personal_finance_management_app/app/app.router.dart';
-import 'package:stacked_services/stacked_services.dart';
+
+import 'app.dart';
 
 void main() {
   Logger.level = Level.verbose;
   setupLocator();
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-      ),
-      navigatorKey: StackedService.navigatorKey,
-      onGenerateRoute: StackedRouter().onGenerateRoute,
-    );
-  }
+  runApp(const PersonalFinanceManagementApp());
 }
